Break out locale data fetch into own middleware
